requirements_mdh.txt
asgiref==3.5.2
boto==2.49.0
boto3==1.10.0
botocore==1.13.50
cachetools==4.2.4
certifi==2022.9.24
chardet==3.0.4
charset-normalizer==2.1.1
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
django-geoposition-2==0.4.0
django-oauth-toolkit==1.2.0
django-oauth2-provider==0.2.6.1
django-requests==0.0.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docutils==0.15.2
et-xmlfile==1.1.0
google-api-core==1.34.0
google-auth==1.35.0
google-cloud==0.34.0
google-cloud-core==1.7.3
google-cloud-translate==1.6.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
grpcio-status==1.48.2
idna==2.7
jmespath==0.10.0
oauthlib==3.2.2
openpyxl==3.0.10
pdfkit==0.6.1
Pillow==9.3.0
protobuf==3.20.3
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
pyshorteners==1.0.1
python-dateutil==2.8.2
python-decouple==3.6
python-http-client==1.2.3
pytz==2022.6
requests==2.20.0
rsa==4.9
s3transfer==0.2.1
sendgrid==2.2.1
shortuuid==1.0.11
simplejwt==2.0.1
six==1.16.0
smtpapi==0.3.1
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
typing==3.7.4.3
urllib3==1.24.3
xlrd==2.0.1
